+++ j2735-2024-ffm-lib/src/main/java/j2735ffm/ConvertException.java
@@ -3,6 +3,7 @@ package j2735ffm;
 /**
  * Checked exception for errors returned from the native library
  */
+@Deprecated(forRemoval = false, since = "3.0.0")
 public class ConvertException extends Exception{
 
   public ConvertException(String message) {